+++ src/components/Home/Home.scss
@@ -9,7 +9,6 @@
 	}
 
 	&-FileInput {
-		font-size: 12px;
 		cursor: pointer;
 
 		&_error {
@@ -53,7 +52,6 @@
 
 		&-Title {
 			text-align: center;
-			font-size: 12px;
 			margin-bottom: 4px;
 			display: block;
 		}
