+++ src/components/shared/theme/components/ThemeButton/ThemeButton.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../scss/variables.scss';
 
 .ThemeButton {
