+++ src/components/shared/theme/components/ThemeButton/ThemeButton.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button, IButtonProps } from '../../library';
+import { Button, IButtonProps } from '../../../library';
 import { cn } from '@bem-react/classname';
 import './ThemeButton.scss';
