+++ src/components/shared/theme/components/ThemeCheckBox/ThemeCheckBox.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../scss/variables.scss';
 
 .ThemeCheckBox {
