+++ src/components/shared/theme/components/ThemeRange/ThemeRange.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../scss/variables.scss';
 
 .ThemeRange {
