+++ src/components/shared/theme/scss/index.scss
@@ -1,0 +1,2 @@
+@import './variables.scss';
+@import './reset.scss';
