+++ src/components/shared/theme/scss/reset.scss
@@ -4,10 +4,9 @@
 	padding: 0;
 	user-select: none;
 	outline: none;
+	font-size: 12px;
 }
 
 body {
-	margin: 0;
-	padding: 0;
 	font-family: sans-serif;
 }
