--- src/scss/index.scss
@@ -1,5 +1,0 @@
-/*
-* Импорт стилей приложения
-*/
-
-@import 'reset.scss';
